requirements.txt
asgiref==3.11.0
dj-database-url==0.5.0
Django==6.0.1
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.2.0
whitenoise==6.9.0
